+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 - termux: use local lsp language servers when mason install not supported
 - termux: disable mason from installing the lua_ls language server
 - termux: neogit v0.0.1 for nvim 0.9.x support
+- termux: conditional check for OS_TERMUX to load config
 
 ## Changed
 - readme: expand on description and use